Use RaceBox for GPS

+++ read_sensors.py
@@ -10,7 +10,7 @@ import time
 import traceback
 import zmq
 
-import read_gps
+import racebox
 
 LOG_DIR = "/home/pi/log/"
 TELEMETRY_DIR = "/home/pi/car_log/"
@@ -34,6 +34,9 @@ longitude = None
 altitude = None
 volts = None
 track = None
+gforce_x = None
+gforce_y = None
+gforce_z = None
 gps_utc_date = None
 gps_utc_time = None
 
@@ -47,15 +50,24 @@ last_log_time = int(time.time())
 next_log_time = last_log_time + 1
 
 context = zmq.Context()
-socket = context.socket(zmq.PUSH)
-socket.connect("tcp://localhost:9961")
+cluster_socket = context.socket(zmq.PUSH)
+racebox_socket = context.socket(zmq.PULL)
+cluster_socket.connect("tcp://localhost:9961")
+racebox_socket.bind("tcp://*:9962")
 
-def send_zmqpp(message):
+# the only other socket should be read-only, but since there are multiple it feels wrong to assume which one we'll be using
+def send_zmqpp(message, socket=cluster_socket):
     if "inf" not in message:
         logger.debug(message)
         socket.send(message.encode())
     else:
         logger.debug("Value is inf; skipping: {}".format(message))
+
+def time_string_from_racebox_data(racebox_data):
+    gps_utc_date = f"{racebox_data['year']}{racebox_data['month']:02}{racebox_data['day']:02}"
+    gps_utc_time = f"{racebox_data['hour']:02}{racebox_data['minute']:02}{racebox_data['second']:02}"
+    return f"{gps_utc_date}-{gps_utc_time}"
+
 
 ttyUSBfiles = []
 ttyUSBs = []
@@ -66,11 +78,6 @@ filenames = os.listdir(serial_device_dir)
 for filename in filenames:
     if "ttyUSB" in filename:
         ttyUSBfiles.append(os.path.join(serial_device_dir, filename))
-# if it's only 1, this will still function without GPS
-if len(ttyUSBfiles) > 2:
-    # I genuinely don't know how to recover from this, so just stop
-    send_zmqpp("STOP\n".encode())
-    sys.exit("Wrong number of USB serial devices: {}".format(len(ttyUSBfiles)))
 for ttyUSBfile in ttyUSBfiles:
     try:
         ttyUSBs.append(serial.Serial(ttyUSBfile, baudrate=921600, timeout=0.5))
@@ -78,32 +85,26 @@ for ttyUSBfile in ttyUSBfiles:
         logger.error(e)
 
 arduino = None
-gps_port = None
 
 logger.debug("Attempting to determine which serial device is which")
-while arduino is None:
-    # Use read() instead of readline() as the two devices will not use the same baudrate and \n will never be found using the wrong baudrate.
-    for ttyUSB in ttyUSBs:
-        try:
-            logger.debug("Reading serial response")
-            data = ttyUSB.read(1024).decode("utf-8")
-            logger.debug("{}: {}".format(ttyUSB.name, data))
-            if data is not None and "FUEL:" in data:
-                logger.debug("Found FUEL: in {}".format(ttyUSB.name))
-                arduino = ttyUSB
-        except AttributeError as e:
-            logger.error(e)
-for ttyUSB in ttyUSBs:
-    if (ttyUSB != arduino):
-        # let the GPS library handle the serial communication
-        gps_port = ttyUSB.name
+if len(ttyUSBs) == 1:
+    arduino = ttyUSB
+# if there are no ttyUSB devices, this block will look for one forever
+elif len(ttyUSBs) > 0:
+    while arduino is None:
+        # Use read() instead of readline() as the two devices will not use the same baudrate and \n will never be found using the wrong baudrate.
+        for ttyUSB in ttyUSBs:
+            try:
+                logger.debug("Reading serial response")
+                data = ttyUSB.read(1024).decode("utf-8")
+                logger.debug("{}: {}".format(ttyUSB.name, data))
+                if data is not None and "FUEL:" in data:
+                    logger.debug("Found FUEL: in {}".format(ttyUSB.name))
+                    arduino = ttyUSB
+            except AttributeError as e:
+                logger.error(e)
+# TODO is this still needed, or was it needed by read_gps?
 time.sleep(0.1)
-
-try:
-    read_gps.debug_enable()
-    read_gps.setup_gps(gps_port)
-except Exception as e:
-    print(traceback.format_exc())
 
 loops = 0
 
@@ -111,11 +112,23 @@ loops = 0
 # TODO using this sort of logging method will not indicate stale data. Use something better.
 LOG_INTERVAL = 0.1
 
-gps_time = read_gps.get_gps_time()
+gps_time = None
 
-logger.info("Read UTC time from GPS ({}). Correcting log filename.".format(gps_time))
-# logging seems to play nicely with this, so it will just continue to log to the new file
-os.rename(log_file_name, os.path.join(LOG_DIR, "read_sensors_{}.log".format(gps_time)))
+first_attempt = time.time()
+while gps_time is None and time.time() - first_attempt < 5:
+    try:
+        racebox_data = racebox_socket.recv_json(flags=zmq.NOBLOCK)
+        if racebox_data["date_time_flags"] & racebox.DATE_TIME_CONFIRM_AVAIL_MASK\
+                and racebox_data["date_time_flags"] & racebox.DATE_VALID_MASK\
+                and racebox_data["date_time_flags"] & racebox.TIME_VALID_MASK:
+            gps_time = time_string_from_racebox_data(racebox_data)
+
+            logger.info("Read UTC time from GPS ({}). Correcting log filename.".format(gps_time))
+            # logging seems to play nicely with this, so it will just continue to log to the new file
+            os.rename(log_file_name, os.path.join(LOG_DIR, "read_sensors_{}.log".format(gps_time)))
+    except zmq.ZMQError:
+        # we'll just continue to use the system time for this file and hope there's not a filename collision
+        pass
 
 # I'm OK with missing a second of data to not log a bunch of repeats to fill the gap between the floor of the timestamp and the actual start time
 next_log_time = int(time.time()) + 1
@@ -141,6 +154,9 @@ with open(output_filename, 'w', newline='') as csvfile:
         "alt",
         "mph",
         "track",
+        "gforce_x",
+        "gforce_y",
+        "gforce_z",
         "repeated",
     ]
     csv_writer = csv.DictWriter(csvfile, fieldnames=fieldnames)
@@ -148,29 +164,33 @@ with open(output_filename, 'w', newline='') as csvfile:
 
     while True:
         try:
-            try:
-                while (gps_data := read_gps.get_position_data(blocking=False)):
-                    logger.debug("Read GPS data ({} message)".format(gps_data["type"]))
-                    if "speed_kn" in gps_data:
-                        mph = gps_data["speed_kn"] * KNOTS_TO_MPH
-                        send_zmqpp("MPH:{}".format(mph))
-                    if "track" in gps_data:
-                        track = gps_data["track"]
-                    if "latitude" in gps_data:
-                        latitude = gps_data["latitude"]
-                    if "longitude" in gps_data:
-                        longitude = gps_data["longitude"]
-                    if "date" in gps_data:
-                        gps_utc_date = gps_data["date"]
-                    if "utc" in gps_data:
-                        gps_utc_time = gps_data["utc"]
-            except AttributeError:
-                pass
+            while True:
+                try:
+                    racebox_data = racebox_socket.recv_json(flags=zmq.NOBLOCK)
+                    if "speed_mph" in racebox_data:
+                        send_zmqpp("MPH:{}".format(racebox_data["speed_mph"]))
+                    if "heading" in racebox_data:
+                        track = racebox_data["heading"]
+                    if "latitude" in racebox_data:
+                        latitude = racebox_data["latitude"]
+                    if "longitude" in racebox_data:
+                        longitude = racebox_data["longitude"]
+                    gps_utc_date = f"{racebox_data['year']}-{racebox_data['month']:02}-{racebox_data['day']:02}"
+                    seconds = racebox_data["second"] + racebox_data["nanoseconds"]/1e9
+                    gps_utc_time = f"{racebox_data['hour']:02}:{racebox_data['minute']:02}:{seconds:05.2}"
+                    gforce_x = racebox_data["gforce_x"]
+                    gforce_y = racebox_data["gforce_y"]
+                    gforce_z = racebox_data["gforce_z"]
+                    volts = racebox_data["input_voltage"]
+                    last_raebox_update = time.time()
+                except zmq.ZMQError:
+                    break
 
             try:
                 while arduino.in_waiting > 0:
                     try:
                         # TODO make this faster. May be an Arduino limitation, but it's slowing down the whole loop considerably (0.05-0.1s)
+                        # splitting it off into yet another script and using more zmq communication may be the answer to accomplish this without moving permanently to async hell
                         message = arduino.readline().decode("utf-8")
                         # messages shouldn't be more than 100 chars. Something seems to be breaking but it's not leaving any error messages.
                         if len(message) > 100:
@@ -200,6 +220,7 @@ with open(output_filename, 'w', newline='') as csvfile:
                         elif name == "FUEL":
                             fuel = float(value)
                             send_zmqpp("FUEL:{}".format(fuel))
+                        # TODO remove this as the RaceBox logs volts. Leaving it in here as a reminder to take it out of the Arduino code.
                         elif name == "VOLTS":
                             volts = float(value)
                             send_zmqpp("VOLTS:{}".format(volts))
@@ -232,6 +253,9 @@ with open(output_filename, 'w', newline='') as csvfile:
                     "alt": altitude,
                     "mph": mph,
                     "track": track,
+                    "gforce_x": gforce_x,
+                    "gforce_y": gforce_y,
+                    "gforce_z": gforce_z,
                     "repeated": 0,
                 }
                 csv_writer.writerow(fields)
